+++ base/math/Random.cpp
@@ -103,7 +103,7 @@ void RandomLegacy::randomize() throw() {
   if (seed == 0) {
     seed = 20010908014640LL;
   }
-  randomize(seed);
+  randomize(static_cast<unsigned int>(seed));
 }
 
 void RandomLegacy::randomize(unsigned int seed) throw()
